+++ packages/widget/src/constants/bridgeable.ts
@@ -436,6 +436,7 @@ export const USDC: BridgeableToken = {
     [CHAINS.BASE.id]: '0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913',
     [CHAINS.POLYGON.id]: '0x2791bca1f2de4661ed88a30c99a7a9449aa84174',
     [CHAINS.SCROLL.id]: '0x06eFdBFf2a14a7c8E15944D1F4A48F9F95F663A4',
+    [CHAINS.LINEA.id]: '0x176211869cA2b568f2A7D4EE941E073a821EE1ff',
   },
   decimals: {
     [CHAINS.ETHEREUM.id]: 6,
@@ -452,6 +453,7 @@ export const USDC: BridgeableToken = {
     [CHAINS.BASE.id]: 6,
     [CHAINS.POLYGON.id]: 6,
     [CHAINS.SCROLL.id]: 6,
+    [CHAINS.LINEA.id]: 6,
   },
   symbol: 'USDC',
   name: 'USD Coin',
@@ -492,6 +494,7 @@ export const USDT: BridgeableToken = {
     [CHAINS.KLAYTN.id]: '0xd6dAb4CfF47dF175349e6e7eE2BF7c40Bb8C05A3',
     [CHAINS.ARBITRUM.id]: '0xfd086bc7cd5c481dcc9c85ebe478a1c0b69fcbb9',
     [CHAINS.AVALANCHE.id]: '0x9702230a8ea53601f5cd2dc00fdbc13d4df4a8c7',
+    [CHAINS.LINEA.id]: '0xA219439258ca9da29E9Cc4cE5596924745e12B93',
   },
   decimals: {
     [CHAINS.ETHEREUM.id]: 6,
@@ -504,6 +507,7 @@ export const USDT: BridgeableToken = {
     [CHAINS.KLAYTN.id]: 6,
     [CHAINS.ARBITRUM.id]: 6,
     [CHAINS.AVALANCHE.id]: 6,
+    [CHAINS.LINEA.id]: 6,
   },
   symbol: 'USDT',
   name: 'USD Tether',
@@ -802,6 +806,7 @@ export const ETH: BridgeableToken = {
     [CHAINS.DFK.id]: '0xfBDF0E31808d0aa7b9509AA6aBC9754E48C58852',
     [CHAINS.BLAST.id]: ZeroAddress,
     [CHAINS.SCROLL.id]: ZeroAddress,
+    [CHAINS.LINEA.id]: ZeroAddress,
   },
   decimals: {
     [CHAINS.ETHEREUM.id]: 18,
@@ -813,6 +818,7 @@ export const ETH: BridgeableToken = {
     [CHAINS.DFK.id]: 18,
     [CHAINS.BLAST.id]: 18,
     [CHAINS.SCROLL.id]: 18,
+    [CHAINS.LINEA.id]: 18,
   },
   symbol: 'ETH',
   name: 'Ethereum',