+++ services/rfq/relayer/relconfig/config_test.go
@@ -21,8 +21,6 @@ func TestChainGetters(t *testing.T) {
 		Chains: map[int]relconfig.ChainConfig{
 			chainID: {
 				RFQAddress:              "0x123",
-				SynapseCCTPAddress:      "0x456",
-				TokenMessengerAddress:   "0x789",
 				Confirmations:           1,
 				NativeToken:             "MATIC",
 				DeadlineBufferSeconds:   10,
@@ -35,12 +33,23 @@ func TestChainGetters(t *testing.T) {
 				QuotePct:                relconfig.NewFloatPtr(0),
 				QuoteWidthBps:           10,
 				QuoteFixedFeeMultiplier: relconfig.NewFloatPtr(1.1),
+				RebalanceConfigs: relconfig.RebalanceConfigs{
+					Synapse: &relconfig.SynapseCCTPRebalanceConfig{
+						SynapseCCTPAddress: "0x456",
+					},
+					Circle: &relconfig.CircleCCTPRebalanceConfig{
+						TokenMessengerAddress: "0x789",
+					},
+					Scroll: &relconfig.ScrollRebalanceConfig{
+						L1GatewayAddress:         "0xabc",
+						L1ScrollMessengerAddress: "0xdef",
+						L2GatewayAddress:         "0xghi",
+					},
+				},
 			},
 		},
 		BaseChainConfig: relconfig.ChainConfig{
 			RFQAddress:              "0x1234",
-			SynapseCCTPAddress:      "0x456",
-			TokenMessengerAddress:   "0x789",
 			Confirmations:           2,
 			NativeToken:             "ARB",
 			DeadlineBufferSeconds:   11,
@@ -53,14 +62,25 @@ func TestChainGetters(t *testing.T) {
 			QuotePct:                relconfig.NewFloatPtr(51),
 			QuoteWidthBps:           11,
 			QuoteFixedFeeMultiplier: relconfig.NewFloatPtr(1.2),
+			RebalanceConfigs: relconfig.RebalanceConfigs{
+				Synapse: &relconfig.SynapseCCTPRebalanceConfig{
+					SynapseCCTPAddress: "0x456",
+				},
+				Circle: &relconfig.CircleCCTPRebalanceConfig{
+					TokenMessengerAddress: "0x789",
+				},
+				Scroll: &relconfig.ScrollRebalanceConfig{
+					L1GatewayAddress:         "0xabc",
+					L1ScrollMessengerAddress: "0xdef",
+					L2GatewayAddress:         "0xghi",
+				},
+			},
 		},
 	}
 	cfg := relconfig.Config{
 		Chains: map[int]relconfig.ChainConfig{
 			chainID: {
 				RFQAddress:              "0x123",
-				SynapseCCTPAddress:      "0x456",
-				TokenMessengerAddress:   "0x789",
 				Confirmations:           1,
 				NativeToken:             "MATIC",
 				DeadlineBufferSeconds:   10,
@@ -80,51 +100,22 @@ func TestChainGetters(t *testing.T) {
 						MaxRebalanceAmount: "1000",
 					},
 				},
+				RebalanceConfigs: relconfig.RebalanceConfigs{
+					Synapse: &relconfig.SynapseCCTPRebalanceConfig{
+						SynapseCCTPAddress: "0x456",
+					},
+					Circle: &relconfig.CircleCCTPRebalanceConfig{
+						TokenMessengerAddress: "0x789",
+					},
+					Scroll: &relconfig.ScrollRebalanceConfig{
+						L1GatewayAddress:         "0xabc",
+						L1ScrollMessengerAddress: "0xdef",
+						L2GatewayAddress:         "0xghi",
+					},
+				},
 			},
 		},
 	}
-
-	t.Run("GetRFQAddress", func(t *testing.T) {
-		defaultVal, err := cfg.GetRFQAddress(badChainID)
-		assert.NoError(t, err)
-		assert.Equal(t, defaultVal, relconfig.DefaultChainConfig.RFQAddress)
-
-		baseVal, err := cfgWithBase.GetRFQAddress(badChainID)
-		assert.NoError(t, err)
-		assert.Equal(t, baseVal, cfgWithBase.BaseChainConfig.RFQAddress)
-
-		chainVal, err := cfgWithBase.GetRFQAddress(chainID)
-		assert.NoError(t, err)
-		assert.Equal(t, chainVal, cfgWithBase.Chains[chainID].RFQAddress)
-	})
-
-	t.Run("GetSynapseCCTPAddress", func(t *testing.T) {
-		defaultVal, err := cfg.GetSynapseCCTPAddress(badChainID)
-		assert.NoError(t, err)
-		assert.Equal(t, defaultVal, relconfig.DefaultChainConfig.SynapseCCTPAddress)
-
-		baseVal, err := cfgWithBase.GetSynapseCCTPAddress(badChainID)
-		assert.NoError(t, err)
-		assert.Equal(t, baseVal, cfgWithBase.BaseChainConfig.SynapseCCTPAddress)
-
-		chainVal, err := cfgWithBase.GetSynapseCCTPAddress(chainID)
-		assert.NoError(t, err)
-		assert.Equal(t, chainVal, cfgWithBase.Chains[chainID].SynapseCCTPAddress)
-	})
-
-	t.Run("GetTokenMessengerAddress", func(t *testing.T) {
-		defaultVal, err := cfg.GetTokenMessengerAddress(badChainID)
-		assert.NoError(t, err)
-		assert.Equal(t, defaultVal, relconfig.DefaultChainConfig.TokenMessengerAddress)
-
-		baseVal, err := cfgWithBase.GetTokenMessengerAddress(badChainID)
-		assert.NoError(t, err)
-		assert.Equal(t, baseVal, cfgWithBase.BaseChainConfig.TokenMessengerAddress)
-
-		chainVal, err := cfgWithBase.GetTokenMessengerAddress(chainID)
-		assert.NoError(t, err)
-		assert.Equal(t, chainVal, cfgWithBase.Chains[chainID].TokenMessengerAddress)
-	})
 
 	t.Run("GetConfirmations", func(t *testing.T) {
 		defaultVal, err := cfg.GetConfirmations(badChainID)
@@ -310,8 +301,6 @@ func TestGetQuoteOffset(t *testing.T) {
 		Chains: map[int]relconfig.ChainConfig{
 			chainID: {
 				RFQAddress:              "0x123",
-				SynapseCCTPAddress:      "0x456",
-				TokenMessengerAddress:   "0x789",
 				Confirmations:           1,
 				NativeToken:             "MATIC",
 				DeadlineBufferSeconds:   10,
@@ -358,7 +347,7 @@ func TestValidation(t *testing.T) {
 						"USDC": {
 							InitialBalancePct:     50,
 							MaintenanceBalancePct: 25,
-							RebalanceMethod:       "synapsecctp",
+							RebalanceMethods:      []string{"synapsecctp"},
 						},
 					},
 				},
@@ -367,7 +356,7 @@ func TestValidation(t *testing.T) {
 						"USDC": {
 							InitialBalancePct:     50,
 							MaintenanceBalancePct: 25,
-							RebalanceMethod:       "synapsecctp",
+							RebalanceMethods:      []string{"synapsecctp"},
 						},
 					},
 				},
@@ -385,7 +374,7 @@ func TestValidation(t *testing.T) {
 						"USDC": {
 							InitialBalancePct:     51,
 							MaintenanceBalancePct: 50,
-							RebalanceMethod:       "synapsecctp",
+							RebalanceMethods:      []string{"synapsecctp"},
 						},
 					},
 				},
@@ -394,7 +383,7 @@ func TestValidation(t *testing.T) {
 						"USDC": {
 							InitialBalancePct:     50,
 							MaintenanceBalancePct: 50,
-							RebalanceMethod:       "synapsecctp",
+							RebalanceMethods:      []string{"synapsecctp"},
 						},
 					},
 				},
@@ -413,7 +402,7 @@ func TestValidation(t *testing.T) {
 						"USDC": {
 							InitialBalancePct:     50,
 							MaintenanceBalancePct: 50,
-							RebalanceMethod:       "synapsecctp",
+							RebalanceMethods:      []string{"synapsecctp"},
 						},
 					},
 				},
@@ -422,7 +411,7 @@ func TestValidation(t *testing.T) {
 						"USDC": {
 							InitialBalancePct:     50,
 							MaintenanceBalancePct: 50.1,
-							RebalanceMethod:       "synapsecctp",
+							RebalanceMethods:      []string{"synapsecctp"},
 						},
 					},
 				},